DDCNL-9071: Remove CompanyCarBenefitFormatters
